BST>Map has my Hashmap in a BST in it. That the code I'm working on
Binary Search Tree using a Map the new ES6 Hashmap instead of just an array and having to deal with collisions and linked lists for expansion or adding and removing. BST doesn't allow duplicates, using strings the BST uses them in alphabetical order. The height of the tree stays short and fast. If I can get rid of the loops then it will be 0(1) I think