first jest test
I'm trying to fill the map with the root which is weaving in the hashmap, and then I'm passing in Key, Value pairs into the BinarySearchTree class.

If we could just call it. I can take it over to repit maybe to test it, then bring it into VS code once it works.

you would just call
console.log(BinarySearchTree("some", "root");